cleanup

+++ src/main/java/com/songoda/ultimateclaims/commands/CommandRound.java
@@ -177,7 +177,7 @@ public class CommandRound extends AbstractCommand {
             plugin.getDataManager().createClaim(claim);
 
             plugin.getLocale().getMessage("command.claim.info")
-                    .processPlaceholder("time", TimeUtils.makeReadable((long) (Settings.STARTING_POWER.getInt() * 60 * 1000)))
+                    .processPlaceholder("time", TimeUtils.makeReadable((long) ((long) Settings.STARTING_POWER.getInt() * 60 * 1000)))
                     .sendPrefixedMessage(sender);
 
         }
